+++ lib/Model/Payment.php
@@ -60,11 +60,11 @@ class Payment implements ModelInterface, ArrayAccess
 'paymentId' => 'string',
 'paymentNumber' => 'string',
 'paymentExternalKey' => 'string',
-'authAmount' => 'BigDecimal',
-'capturedAmount' => 'BigDecimal',
-'purchasedAmount' => 'BigDecimal',
-'refundedAmount' => 'BigDecimal',
-'creditedAmount' => 'BigDecimal',
+'authAmount' => 'float',
+'capturedAmount' => 'float',
+'purchasedAmount' => 'float',
+'refundedAmount' => 'float',
+'creditedAmount' => 'float',
 'currency' => 'string',
 'paymentMethodId' => 'string',
 'transactions' => '\Killbill\Client\Swagger\Model\PaymentTransaction[]',
@@ -720,7 +720,7 @@ self::CURRENCY_BTC,        ];
     /**
      * Gets authAmount
      *
-     * @return BigDecimal
+     * @return float
      */
     public function getAuthAmount()
     {
@@ -730,7 +730,7 @@ self::CURRENCY_BTC,        ];
     /**
      * Sets authAmount
      *
-     * @param BigDecimal $authAmount authAmount
+     * @param float $authAmount authAmount
      *
      * @return $this
      */
@@ -744,7 +744,7 @@ self::CURRENCY_BTC,        ];
     /**
      * Gets capturedAmount
      *
-     * @return BigDecimal
+     * @return float
      */
     public function getCapturedAmount()
     {
@@ -754,7 +754,7 @@ self::CURRENCY_BTC,        ];
     /**
      * Sets capturedAmount
      *
-     * @param BigDecimal $capturedAmount capturedAmount
+     * @param float $capturedAmount capturedAmount
      *
      * @return $this
      */
@@ -768,7 +768,7 @@ self::CURRENCY_BTC,        ];
     /**
      * Gets purchasedAmount
      *
-     * @return BigDecimal
+     * @return float
      */
     public function getPurchasedAmount()
     {
@@ -778,7 +778,7 @@ self::CURRENCY_BTC,        ];
     /**
      * Sets purchasedAmount
      *
-     * @param BigDecimal $purchasedAmount purchasedAmount
+     * @param float $purchasedAmount purchasedAmount
      *
      * @return $this
      */
@@ -792,7 +792,7 @@ self::CURRENCY_BTC,        ];
     /**
      * Gets refundedAmount
      *
-     * @return BigDecimal
+     * @return float
      */
     public function getRefundedAmount()
     {
@@ -802,7 +802,7 @@ self::CURRENCY_BTC,        ];
     /**
      * Sets refundedAmount
      *
-     * @param BigDecimal $refundedAmount refundedAmount
+     * @param float $refundedAmount refundedAmount
      *
      * @return $this
      */
@@ -816,7 +816,7 @@ self::CURRENCY_BTC,        ];
     /**
      * Gets creditedAmount
      *
-     * @return BigDecimal
+     * @return float
      */
     public function getCreditedAmount()
     {
@@ -826,7 +826,7 @@ self::CURRENCY_BTC,        ];
     /**
      * Sets creditedAmount
      *
-     * @param BigDecimal $creditedAmount creditedAmount
+     * @param float $creditedAmount creditedAmount
      *
      * @return $this
      */
